collection/requirements.txt
azure-functions
azure-storage-blob
pandas
pyarrow
requests
